+++ src/utils/api_utils.ts
@@ -15,29 +15,6 @@ export const checkServerDown = async () => {
 };
 
 
-export const getUserForToken = async (accessToken: string) => {
-	let isTokenValid = false;
-	const response = <any> await fetch(
-		API_ROUTES.USERS, {
-		headers: {
-			'Content-Type': 'application/json',
-			'Authorization': `Basic ${accessToken}`
-		},
-	})
-	.then(res => res.json())
-	.then(json => json)
-	.catch(err => {
-		isTokenValid = false;
-	});
-	if (response && !(response.error)) {
-		isTokenValid = true;
-	}
-	return {
-		isTokenValid,
-		response
-	};
-};
-
 export const createUserWithApi = async (accessToken: string) => {
 	let error = "";
 	let email = "";