+++ tests/test_init/handler.test.js
@@ -19,7 +19,8 @@ import {
     TEST_USER,
     waitFor,
     addUser,
-    writeTestRepoFiles
+    writeTestRepoFiles,
+    TEST_EMAIL
 } from "../helpers/helpers";
 import { SYNC_IGNORE_FILE_DATA } from "../../src/constants";
 import { pathUtils } from "../../src/utils/path_utils";
@@ -28,11 +29,12 @@ import { createSystemDirectories } from "../../src/utils/setup_utils";
 import { CodeSyncState, CODESYNC_STATES } from "../../src/utils/state_utils";
 import { s3UploaderUtils } from "../../src/init/s3_uploader";
 
+
 describe("initHandler: connectRepo", () => {
     let baseRepoPath;
     let repoPath;
     let configPath;
-    const user = { email: "dummy@email.cpm"};
+    const userResponse = {user: {email: "dummy@email.cpm"}};
 
     beforeEach(() => {
         jest.clearAllMocks();
@@ -60,7 +62,7 @@ describe("initHandler: connectRepo", () => {
 
     test("Server is down", async () => {
         fetchMock.mockResponseOnce(null);
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         // Verify error msg
         expect(vscode.window.showErrorMessage).toHaveBeenCalledTimes(1);
@@ -70,8 +72,8 @@ describe("initHandler: connectRepo", () => {
     test("Invalid access token", async () => {
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(INVALID_TOKEN_JSON));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+            .mockResponseOnce(JSON.stringify(INVALID_TOKEN_JSON), { status: 401 });
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         // Verify error msg
         expect(vscode.window.showErrorMessage).toHaveBeenCalledTimes(1);
@@ -83,8 +85,8 @@ describe("initHandler: connectRepo", () => {
     test("Repo already synced", async () => {
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+            .mockResponseOnce(JSON.stringify(userResponse));
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         // Add repo in config
         const configUtil = new Config(repoPath, configPath);
         configUtil.addRepo();
@@ -99,12 +101,12 @@ describe("initHandler: connectRepo", () => {
         const syncIgnorePath = path.join(repoPath, SYNCIGNORE);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
+            .mockResponseOnce(JSON.stringify(userResponse));
         // Add repo in config
         const configUtil = new Config(repoPath, configPath);
         configUtil.addRepo(true);
         addUser(baseRepoPath);
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         // Verify error msg
         expect(vscode.window.showWarningMessage).toHaveBeenCalledTimes(0);
@@ -117,10 +119,10 @@ describe("initHandler: connectRepo", () => {
         const subDir = path.join(repoPath, "directory");
         fs.mkdirSync(subDir);
         const syncIgnorePath = path.join(subDir, SYNCIGNORE);
-        const handler = new initHandler(subDir, "ACCESS_TOKEN");
+        const handler = new initHandler(subDir, "ACCESS_TOKEN", TEST_EMAIL);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
+            .mockResponseOnce(JSON.stringify(userResponse));
         // Add repo in config
         const configUtil = new Config(repoPath, configPath);
         configUtil.addRepo(true);
@@ -138,8 +140,8 @@ describe("initHandler: connectRepo", () => {
     test(".syncignore should be created", async () => {
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+            .mockResponseOnce(JSON.stringify(userResponse));
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         // Verify error msg
         const syncIgnorePath = path.join(repoPath, SYNCIGNORE);
@@ -157,8 +159,8 @@ describe("initHandler: connectRepo", () => {
         fs.writeFileSync(gitignorePath, gitignoreData);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+            .mockResponseOnce(JSON.stringify(userResponse));
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         // Verify error msg
         const syncIgnorePath = path.join(repoPath, SYNCIGNORE);
@@ -176,8 +178,8 @@ describe("initHandler: connectRepo", () => {
         fs.writeFileSync(syncIgnorePath, syncIgnoreData);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN");
+            .mockResponseOnce(JSON.stringify(userResponse));
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL);
         await handler.connectRepo();
         expect(fs.existsSync(syncIgnorePath)).toBe(true);
         const _syncIgnoreData = readFile(syncIgnorePath);
@@ -193,8 +195,8 @@ describe("initHandler: connectRepo", () => {
         fs.writeFileSync(syncIgnorePath, syncIgnoreData);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN", true);
+            .mockResponseOnce(JSON.stringify(userResponse));
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL, true);
         await handler.connectRepo();
         // Verify .syncignore has been created
         expect(fs.existsSync(syncIgnorePath)).toBe(true);
@@ -216,7 +218,7 @@ describe("initHandler: Syncing Branch", () => {
     let pathUtilsObj;
     let shadowFilePath;
     let originalsFilePath;
-    const user = {email: "dummy@email.cpm"};
+    const userResponse = {user: {email: "dummy@email.cpm"}};
 
     beforeEach(() => {
         jest.clearAllMocks();
@@ -247,9 +249,9 @@ describe("initHandler: Syncing Branch", () => {
         fs.writeFileSync(filePath, DUMMY_FILE_CONTENT);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user))
+            .mockResponseOnce(JSON.stringify(userResponse))
             .mockResponseOnce(JSON.stringify({ status: false }));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN", true);
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL, true);
         await handler.connectRepo();
         // Verify error msg
         expect(vscode.window.showErrorMessage).toHaveBeenCalledTimes(0);
@@ -269,10 +271,10 @@ describe("initHandler: Syncing Branch", () => {
         fs.writeFileSync(filePath, DUMMY_FILE_CONTENT);
         fetchMock
             .mockResponseOnce(JSON.stringify({ status: true }))
-            .mockResponseOnce(JSON.stringify(user))
+            .mockResponseOnce(JSON.stringify(userResponse))
             .mockResponseOnce(JSON.stringify({ status: true }))
             .mockResponseOnce(JSON.stringify(TEST_REPO_RESPONSE));
-        const handler = new initHandler(repoPath, "ACCESS_TOKEN", true);
+        const handler = new initHandler(repoPath, "ACCESS_TOKEN", TEST_EMAIL, true);
         await handler.connectRepo();
         // Verify error msg
         expect(vscode.window.showErrorMessage).toHaveBeenCalledTimes(0);